+++ editor/src/components/canvas/canvas-strategies/strategies/grid-helpers.ts
@@ -777,3 +777,53 @@ export function removeGridTemplateDimensionAtIndex(
     },
   })
 }
+
+// Return an array of related indexes to a given index inside a grid's template dimensions.
+export function getGridRelatedIndexes(params: {
+  template: GridDimension[]
+  index: number
+}): number[] {
+  let relatedIndexes: number[][][] = [] // This looks scary but it's not! It's just a list of indexes, containing a list of the indexes *per group element*.
+  // For example, 1fr repeat(3, 10px 20px) 1fr, will be represented as:
+  /**
+   * [
+   *  [ [0] ]
+   *  [ [1, 3] [2, 4]  ]
+   *  [ [5] ]
+   * ]
+   */
+  let elementCount = 0 // basically the expanded index
+  for (const dim of params.template) {
+    if (dim.type === 'REPEAT') {
+      let groupIndexes: number[][] = []
+      // for each value push the related indexes as many times as the repeats counter
+      for (let valueIndex = 0; valueIndex < dim.value.length; valueIndex++) {
+        let repeatedValueIndexes: number[] = []
+        for (let repeatIndex = 0; repeatIndex < dim.times; repeatIndex++) {
+          repeatedValueIndexes.push(elementCount + valueIndex + repeatIndex * dim.value.length)
+        }
+        groupIndexes.push(repeatedValueIndexes)
+      }
+      relatedIndexes.push(groupIndexes)
+      elementCount += dim.value.length * dim.times // advance the counter as many times as the repeated values *combined*
+    } else {
+      relatedIndexes.push([[elementCount]])
+      elementCount++
+    }
+  }
+
+  // Now, expand the indexes calculated above so they "flatten out" to match the generated values
+  let expandedRelatedIndexes: number[][] = []
+  params.template.forEach((dim, dimIndex) => {
+    if (dim.type === 'REPEAT') {
+      for (let repeatIndex = 0; repeatIndex < dim.times * dim.value.length; repeatIndex++) {
+        const indexes = relatedIndexes[dimIndex][repeatIndex % dim.value.length]
+        expandedRelatedIndexes.push(indexes)
+      }
+    } else {
+      expandedRelatedIndexes.push(relatedIndexes[dimIndex][0])
+    }
+  })
+
+  return expandedRelatedIndexes[params.index] ?? []
+}
